# use regexp for speed up